基本功能可用

+++ src/main/java/model/User.java
@@ -47,10 +47,10 @@ public class User implements Serializable {
     private Boolean isVip;
 
     @Column(name = "memory_size", nullable = false)
-    private Integer memorySize;
+    private Long memorySize;
 
     @Column(name = "used_size", nullable = false)
-    private Integer usedSize;
+    private Long usedSize;
 
     @Column(name = "private_status", nullable = false)
     private Boolean privateStatus;
@@ -168,19 +168,19 @@ public class User implements Serializable {
         this.isVip = isVip;
     }
 
-    public Integer getMemorySize() {
+    public Long getMemorySize() {
         return memorySize;
     }
 
-    public void setMemorySize(Integer memorySize) {
+    public void setMemorySize(Long memorySize) {
         this.memorySize = memorySize;
     }
 
-    public Integer getUsedSize() {
+    public Long getUsedSize() {
         return usedSize;
     }
 
-    public void setUsedSize(Integer usedSize) {
+    public void setUsedSize(Long usedSize) {
         this.usedSize = usedSize;
     }
